+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `pdfConfigurations` to email submission event validation
+
 ## [12.1.0] - 2025-09-15
 
 ### Added